Move spec_list within spack.environment
