+++ lib/spack/spack/environment/list.py
